requirements.txt
streamlit==1.37.1
pytesseract==0.3.10
opencv-python-headless==4.10.0.84
Pillow==10.4.0
pdf2image==1.17.0
pypdf==4.3.1
groq==0.9.0
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5
python-dotenv==1.0.1
PyMuPDF==1.24.7